🚚 rename: BonusNumber 클래스 파일명 수정
